requirements.txt
annotated-types==0.5.0
apiclient==1.0.4
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
Flask==3.0.2
frozenlist==1.4.0
future==0.18.2
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.11
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyGithub==2.2.0
PyGObject==3.42.1
pygsheets==2.0.6
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==3.0.1
wrapt==1.16.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
